xwayland: correct mouse position and better logic for mode settig detection
mouse: correct mouse input when a viewport is set. wp_viewporter does not
do that itself.

logic: try to have a somewhat reasonable logic to detect which window tried
to set a mode. Currently a randr set-mode-call doesn't change anything,
neither do we know which client requested it.
So if a mode is set, we save the requested dimensions and check
whether the current focus window has an equivalent size.
Also, if a window changes it size, we check if there's currently
a corresponding mode set.
We could further restrict false positives by setting a time stamp and
only set the viewport if a window resize and a mode set was almost
simultanious.